pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastapi>=0.120.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
    "sse-starlette>=3.0.2",
    "uvicorn[standard]>=0.38.0",
]
